% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{all_boards}
\alias{all_boards}
\title{All Boards}
\usage{
all_boards(...)
}
\arguments{
\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Retrieves a named list of all boards and information associated with them.
}
\keyword{internal}
