% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_board.R
\name{auto_board}
\alias{auto_board}
\title{Create a board from a url}
\usage{
auto_board(url)
}
\arguments{
\item{url}{A url to e.g. a vanity_url on a Connect server}
}
\value{
the board created from the url
}
\description{
Create a board from a url
}
