% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_auto.R
\name{board_auto}
\alias{board_auto}
\title{Create an automagic board around a single (vanity) url}
\usage{
board_auto(url)
}
\arguments{
\item{url}{the url}
}
\value{
the board
}
\description{
Create an automagic board around a single (vanity) url
}
