% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_azure.R
\name{board_azure}
\alias{board_azure}
\title{Use an Azure Blob Storage Container as a board}
\usage{
board_azure(
  account,
  container,
  versioned = TRUE,
  key = NULL,
  token = NULL,
  sas = NULL,
  url = NULL,
  cache = NULL
)
}
\arguments{
\item{account, container}{Account and container name.}

\item{versioned}{Should this board be registered with support for versions?}

\item{key, token, sas}{Authentication credentials: either an access \code{key}, an
Azure Active Directory (AAD) \code{token}, or a \code{SAS}, in that order of
priority. If no omitted, provides read access to a public (anonymous) board.}

\item{url}{If omitted, \code{board_azure()} will generate a endpoint url of the
form \code{"https://{account}.blob.core.windows.net/{container}"}. Use this argument
to override if needed.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}
}
\description{
Pin data to a container on Azure's blog storage using the AzureStor package.
}
\examples{
if (requireNamespace("AzureStor")) {
  # Public access board
  board <- board_azure("pins", "public-data")
  board \%>\% pin_read("mtcars")
}

\dontrun{
# To create a board that you can write to, you'll need to supply one
# of `key`, `token`, or `sas`
board <- board_azure("pins", "private-data", sas = "...")
board \%>\% pin_write(iris)
}
}
