% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_cache_path}
\alias{board_cache_path}
\title{Retrieve Default Cache Path}
\usage{
board_cache_path(name)
}
\arguments{
\item{name}{Board name}
}
\description{
Retrieves the default path used to cache boards and pins. Makes
use of the \code{rappdirs} package to use cache folders
defined by each OS.
}
\examples{
# retrieve default cache path
board_cache_path("local")
}
\keyword{internal}
