% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{board_create_pin}
\alias{board_create_pin}
\title{Extensible API}
\usage{
board_create_pin(board, x, name, description, type, metadata)
}
\arguments{
\item{x}{A local file path or an object. Boards must support storing both.}
}
\description{
Family of functions meant to be used to extend pins to support new
boards, not to be used by end users.
}
\keyword{internal}
