% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_default}
\alias{board_default}
\title{Default Board}
\usage{
board_default()
}
\description{
Retrieves the default board, which defaults to \code{"temp"} but can also be
configured with the \code{pins.board} option.
}
\examples{

# configure default board
options(pind.board = "temp")

# retrieve default board
board_default()

}
