% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_deparse}
\alias{board_deparse}
\title{Deparse a board object}
\usage{
board_deparse(board, ...)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_url]{board_url()}} or other other \code{board_} function.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\description{
Create a character string of R code for a board object. Board
characteristics like a URL, server, region, etc. are included but
authentication details are \emph{not} included. This function is useful for
generating code to access a specific board from a remote server.
}
\examples{
\dontrun{
board <- board_rsconnect()
# Generate code to access this board from elsewhere
board_deparse(board)
}

}
