% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_disconnect}
\alias{board_disconnect}
\title{Disconnect to Board}
\usage{
board_disconnect(name, ...)
}
\arguments{
\item{name}{The name of the board to deactivate.}

\item{...}{Additional parameters required to disconnect from a particular board.}
}
\description{
Disconnects board from RStudio's connection pane, when available.
}
\keyword{internal}
