% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_dospaces.R
\name{board_dospace}
\alias{board_dospace}
\title{Register DigitalOcean Board}
\usage{
board_dospace(
  name = "dospace",
  space = Sys.getenv("DO_SPACE"),
  key = Sys.getenv("DO_ACCESS_KEY_ID"),
  secret = Sys.getenv("DO_SECRET_ACCESS_KEY"),
  datacenter = Sys.getenv("DO_DATACENTER"),
  cache = NULL,
  host = "digitaloceanspaces.com",
  ...
)
}
\arguments{
\item{name}{Board name, used to identify board in actions that affect
multiple boards.}

\item{space}{The name of the DigitalOcean space.}

\item{key, secret}{The key and secret for your space. You can create
a key and secret in the "Spaces access keys" in your API settings.

The \code{secret} is equivalent to a password, so generally should not be stored
in your script. The easiest alternative is to store it in the
\code{AZURE_STORAGE_KEY} environment variable, which \code{board_azure()} will
use by default.}

\item{datacenter}{The datacenter name.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{host}{The host to use for storage, defaults to \code{"digitaloceanspaces.com"}.}

\item{...}{Additional parameters stored in the board object.}
}
\description{
To use DigitalOcean Spaces as a \href{boards-understanding.html}{board}, you first
need an DigitalOcean space and a storage key. You can sign-up and create
those at \href{https://www.digitalocean.com/}{digitalocean.com}.
}
\examples{
\dontrun{
# the following example requires a DigitalOcean Spaces API key
board <- board_dospace(bucket = "s3bucket")
}
}
\keyword{internal}
