% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_local.R
\name{board_folder}
\alias{board_folder}
\alias{board_local}
\alias{board_temp}
\title{Use a local folder as board}
\usage{
board_folder(path, name = "folder", versions = FALSE)

board_local(versions = FALSE)

board_temp(name = "temp", versions = FALSE)
}
\arguments{
\item{path}{Path to directory to store pins. Will be created if it
doesn't already exist.}

\item{name}{An optional name used identify the board. This is no longer
generally needed since you should be passing around an explicit
board object.}

\item{versions}{Should this board be registered with support for versions?}
}
\description{
\itemize{
\item \code{board_folder()} creates a board inside a folder. You can use this to
share files by using a folder on a shared network drive or inside
a DropBox.
\item \code{board_temp()} creates a temporary board that lives in a session
specific temporary directory. It will be automatically deleted once
the current R session ends.
\item \code{board_local()} is mostly included for backward compatiblity. It's
the default board used by \code{\link[=pin]{pin()}} and \code{\link[=pin_get]{pin_get()}}.
}
}
\examples{
# session-specific local board
board <- board_temp()
}
\seealso{
Other boards: 
\code{\link{board_azure}()},
\code{\link{board_github}()},
\code{\link{board_rsconnect}()},
\code{\link{board_s3}()}
}
\concept{boards}
