% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_kaggle.R
\name{board_kaggle_competitions}
\alias{board_kaggle_competitions}
\title{Retrieve data from kaggle competition}
\usage{
board_kaggle_competitions(username = NULL, key = NULL, cache = NULL)
}
\arguments{
\item{username, key}{Your kaggle username and key, if you choose not to use
\code{"~/.kaggle/kaggle.json"}.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}
}
\description{
\itemize{
\item Data is only re-downloaded when it changes.
\item Auth with \code{"~/.kaggle/kaggle.json"}
\item Will almost always need to use \code{pin_download()} rather than \code{pin_read()}
}
}
\examples{
board <- board_kaggle_competitions()
board

board \%>\% pin_meta("titanic")
paths <- board \%>\% pin_download("titanic")
paths
head(read.csv(paths[[1]]))
head(read.csv(paths[[2]]))
}
