% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_ms365.R
\name{board_ms365}
\alias{board_ms365}
\alias{board_sharepoint}
\alias{board_onedrive}
\title{Use a OneDrive or Sharepoint document library as a board}
\usage{
board_ms365(drive, path, versioned = TRUE, cache = NULL)

board_sharepoint(drive, path, versioned = TRUE, cache = NULL)

board_onedrive(drive, path, versioned = TRUE, cache = NULL)
}
\arguments{
\item{drive}{A OneDrive or SharePoint document library object.}

\item{versioned}{Should this board be registered with support for versions?}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}
}
\description{
Pin data to a folder in Onedrive or a SharePoint Online document library
using the Microsoft365R package.
}
\examples{
if (requireNamespace("AzureStor")) {
  # Public access board
  url <- "https://pins.blob.core.windows.net/public-data"
  container <- AzureStor::blob_container(url)
  board <- board_azure(container)
  board \%>\% pin_read("mtcars")
}

\dontrun{
# To create a board that you can write to, you'll need to supply one
# of `key`, `token`, or `sas` to AzureStor::blob_container()
container <- AzureStor::blob_container(url, key = "my-key")
board <- board_azure(container)
board \%>\% pin_write(iris)
}
}
