% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_azure.R
\name{board_register_azure}
\alias{board_register_azure}
\alias{board_azure}
\title{Register Azure Board}
\usage{
board_register_azure(
  name = "azure",
  container = Sys.getenv("AZURE_STORAGE_CONTAINER"),
  account = Sys.getenv("AZURE_STORAGE_ACCOUNT"),
  key = Sys.getenv("AZURE_STORAGE_KEY"),
  cache = board_cache_path(),
  path = NULL,
  ...
)

board_azure(
  name = "azure",
  container = Sys.getenv("AZURE_STORAGE_CONTAINER"),
  account = Sys.getenv("AZURE_STORAGE_ACCOUNT"),
  key = Sys.getenv("AZURE_STORAGE_KEY"),
  cache = NULL,
  ...
)
}
\arguments{
\item{name}{Board name, used to identify board in actions that affect
multiple boards.}

\item{container}{The name of the Azure Storage container.}

\item{account}{The name of the Azure Storage account.}

\item{key}{The access key for the Azure Storage container. You can find
this under "Access keys" in your storage account settings.

The \code{key} is equivalent to a password, so generally should not be stored
in your script. The easiest alternative is to store it in the
\code{AZURE_STORAGE_KEY} environment variable, which \code{board_azure()} will
use by default.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{path}{Subdirectory within \code{url}}

\item{...}{Additional parameters stored in the board object.}
}
\description{
To use Microsoft Azure Storage as a board, you'll need an Azure Storage
account, an Azure Storage container, and an Azure Storage key.
You can sign-up and create those at \href{https://portal.azure.com}{portal.azure.com}.
}
\examples{
\dontrun{
# the following example requires an Azure Storage key
board_register_azure(container = "pinscontainer",
                     account = "pinsstorage",
                     key = "abcabcabcabcabcabcabcabcabcab==")
}
}
\seealso{
board_register

Other boards: 
\code{\link{board_register_github}()},
\code{\link{board_register_local}()},
\code{\link{board_register_rsconnect}()},
\code{\link{board_register_s3}()}
}
\concept{boards}
