% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_datatxt.R
\name{board_register_datatxt}
\alias{board_register_datatxt}
\alias{board_datatxt}
\title{Register \code{data.txt} Board}
\usage{
board_register_datatxt(
  url,
  name = NULL,
  headers = NULL,
  cache = board_cache_path(),
  ...
)

board_datatxt(
  url,
  name = NULL,
  headers = NULL,
  cache = board_cache_path(),
  needs_index = TRUE,
  browse_url = url,
  index_updated = NULL,
  index_randomize = FALSE,
  path = NULL,
  versions = FALSE,
  ...
)
}
\arguments{
\item{url}{Path to the \code{data.txt} file or directory containing it.}

\item{name}{Board name, used to identify board in actions that affect
multiple boards.}

\item{headers}{Optional list of headers to include or a function to generate them.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{...}{Additional parameters stored in the board object.}

\item{needs_index}{Does this board have an index file?}

\item{browse_url}{Not currently used}

\item{index_updated}{Callback function used to update index}

\item{index_randomize}{When retrieving \code{data.txt} at a parameter with random
query string to defeat caching?}

\item{path}{Subdirectory within \code{url}}

\item{versions}{Should this board be registered with support for versions?}
}
\description{
Register a board that for a website that uses the \href{https://datatxt.org}{data.txt}
specification. A \code{data.txt} file is a YAML that provides some basic metadata
about a directory of files.
}
\examples{

# register website board using datatxt file
board_register_datatxt(
  url = "https://datatxt.org/data.txt",
  name = "txtexample",
  cache = tempfile()
)

# find pins
pin_find(board = "txtexample")

}
\seealso{
board_register
}
