% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_dospaces.R
\name{board_register_dospace}
\alias{board_register_dospace}
\alias{board_dospace}
\title{Register DigitalOcean Board}
\usage{
board_register_dospace(
  name = "dospace",
  space = Sys.getenv("DO_SPACE"),
  key = Sys.getenv("DO_ACCESS_KEY_ID"),
  secret = Sys.getenv("DO_SECRET_ACCESS_KEY"),
  datacenter = Sys.getenv("DO_DATACENTER"),
  cache = board_cache_path(),
  host = "digitaloceanspaces.com",
  path = NULL,
  ...
)

board_dospace(
  name = "dospace",
  space = Sys.getenv("DO_SPACE"),
  key = Sys.getenv("DO_ACCESS_KEY_ID"),
  secret = Sys.getenv("DO_SECRET_ACCESS_KEY"),
  datacenter = Sys.getenv("DO_DATACENTER"),
  cache = NULL,
  host = "digitaloceanspaces.com",
  ...
)
}
\arguments{
\item{name}{Board name, used to identify board in actions that affect
multiple boards.}

\item{space}{The name of the DigitalOcean space. Defaults to the \code{DO_SPACE} environment
variable.}

\item{key}{The key of the DigitalOcean space. Defaults to the \code{DO_ACCESS_KEY_ID} environment
variable.}

\item{secret}{The secret of the DigitalOcean space. Defaults to the \code{DO_SECRET_ACCESS_KEY} environment
variable.}

\item{datacenter}{The datacenter of the DigitalOcean space. Defaults to the \code{DO_DATACENTER} environment
variable.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{host}{The host to use for storage, defaults to \code{"digitaloceanspaces.com"}.}

\item{path}{Subdirectory within \code{url}}

\item{...}{Additional parameters stored in the board object.}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register a DigitalOcean Spaces board.
}
\details{
This function requires a DigitalOcean space to be manually created; otherwise,
registering a DigitalOcean space will fail.
}
\examples{
\dontrun{
# the following example requires a DigitalOcean Spaces API key
board_register_s3(bucket = "s3bucket")
}
}
\seealso{
board_register
}
