% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_gcloud.R
\name{board_register_gcloud}
\alias{board_register_gcloud}
\alias{board_gcloud}
\title{Register Google Cloud Board}
\usage{
board_register_gcloud(
  name = "gcloud",
  bucket = Sys.getenv("GCLOUD_STORAGE_BUCKET"),
  token = NULL,
  cache = board_cache_path(),
  path = NULL,
  ...
)

board_gcloud(
  name,
  bucket = Sys.getenv("GCLOUD_STORAGE_BUCKET"),
  token = NULL,
  cache = NULL,
  ...
)
}
\arguments{
\item{name}{Board name, used to identify board in actions that affect
multiple boards.}

\item{bucket}{The name of the Google Cloud Storage bucket. Defaults to the \code{GCLOUD_STORAGE_BUCKET} environment
variable.}

\item{token}{The access token of the Google Cloud Storage container.
Generally, it's best to leave this as \code{NULL}, and rely on the installed
Google Cloud SDK to handle authentication.

If you do want to use an access token, you can retrieve it from
\url{https://developers.google.com/oauthplayground}. You will need to
authorize the "Google Storage API v1" scope.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{path}{Subdirectory within \code{url}}

\item{...}{Additional parameters stored in the board object.}
}
\description{
To use a Google Cloud Storage board, you first need a Google Cloud Storage
account, a Google Storage bucket, and an access token or the
\href{https://cloud.google.com/sdk/}{Google Cloud SDK} properly installed and
configured. You can sign-up and create these from
\url{https://console.cloud.google.com}
}
\examples{
\dontrun{
# the following example requires the Google Cloud SDK to be configured
board_register_gcloud(container = "gcloudcontainer")
}
}
\seealso{
board_register
}
