% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_github.R
\name{board_register_github}
\alias{board_register_github}
\alias{board_github}
\title{Register GitHub Board}
\usage{
board_register_github(
  name = "github",
  repo = NULL,
  branch = NULL,
  token = NULL,
  path = "",
  host = "https://api.github.com",
  cache = board_cache_path(),
  ...
)

board_github(
  name,
  repo = NULL,
  branch = NULL,
  token = NULL,
  path = "",
  host = "https://api.github.com",
  ...
)
}
\arguments{
\item{name}{Optional name for this board, defaults to 'github'.}

\item{repo}{The GitHub repository formatted as 'owner/repo', can be
\code{NULL} if the \code{GITHUB_PAT} environment variable is set.}

\item{branch}{The branch to use to commit pins.}

\item{token}{Token to use when \code{GITHUB_PAT} is not specified.}

\item{path}{The subdirectory in the repo where the pins will be stored.}

\item{host}{The URL of the GitHub API. You'll need to customise
this to use GitHub enterprise, e.g. \code{"https://yourhostname/api/v3"}.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
To use a GitHub board, you'll need to set up authentication. This is likely
to just work if you already use GitHub for other purposes since pins will
inspect your  \code{GITHUB_PAT} environment variable. Otherwise, you can follow
the instructions at \url{https://happygitwithr.com/credential-caching.html} to
get set up.
}
\section{Large Files}{
A GitHub repo only supports files under 25MB in size (100MB in theory but
there is additional overhead when using the GitHub API). To store large
files, GitHub
\href{https://help.github.com/en/articles/distributing-large-binaries}{recommends}
storing them using GitHub Releases which support up to 2GB files, which
is what pins uses. You don't need to do anything extra as this will happen
behind the scenes, but don't be surprised if pins creates releases in your
repo.
}

\examples{
\dontrun{
# the following example requires a GitHub API key
board_register_github(repo = "owner/repo")
}
}
\seealso{
board_register

Other boards: 
\code{\link{board_register_azure}()},
\code{\link{board_register_local}()},
\code{\link{board_register_rsconnect}()},
\code{\link{board_register_s3}()}
}
\concept{boards}
