% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registration.R
\name{board_register_s3}
\alias{board_register_s3}
\title{Register S3 Board}
\usage{
board_register_s3(name = "github", bucket, cache = board_cache_path())
}
\arguments{
\item{name}{Optional name for this board, defaults to 'github'.}

\item{bucket}{The name of the Amazon S3 bucket.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register an Amazon S3 bucket as a board.
}
\details{
This function requires an Amazon S3 bucket to be manually created; otherwise,
registering an S3 board will fail.
}
\examples{
\dontrun{
# the following example requires an Amazon S3 API key
board_register_s3(bucket = "owner/repo")
}
}
\seealso{
board_register
}
