% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_unregister}
\alias{board_unregister}
\title{Unregister Board}
\usage{
board_unregister(name)
}
\arguments{
\item{name}{An optional name to identify this board, defaults to the board name.}
}
\description{
Unregisters a board, useful to disable boards no longer in use.
}
