% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connection}
\alias{connection}
\title{Describe Connection}
\usage{
connection(initializer = "DBI::dbConnect", driver = NULL, ...)
}
\arguments{
\item{initializer}{Function used to connect. Defaults to \code{"DBI::dbConnect"}.}

\item{driver}{The name of the driver with an optional library}

\item{...}{Additional parameters to initialize the connection.}
}
\description{
Describe a connection to be able to \code{pin()} it.
}
