% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crandatasets}
\alias{crandatasets}
\title{Index of datasets found in CRAN packages}
\format{An object of class \code{"data.frame"}; see \code{\link[qtl]{data.frame}}.}
\source{
\href{https://cran.rstudio.com/}{CRAN}
}
\usage{
data(crandatasets)
}
\description{
A dataset containing an index of all CRAN datasets. Contains a package, dataset name
and description.
}
\examples{
data(crandatasets)
}
\keyword{datasets}
