% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cranfiles}
\alias{cranfiles}
\title{Index of tabular files found in CRAN packages}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://cran.r-project.org}{CRAN}
}
\usage{
data(cranfiles)
}
\description{
A dataset containing an index of all CRAN files with tabular data. Contains a package,
dataset name and description.
}
\examples{
data(cranfiles)
}
\keyword{datasets}
\keyword{internal}
