% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_local_storage}
\alias{board_local_storage}
\title{Local storage path}
\usage{
board_local_storage(...)
}
\description{
Deprecated: please use \code{\link[=board_cache_path]{board_cache_path()}} instead.
}
\keyword{internal}
