% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_extensions.R
\name{board_create_pin}
\alias{board_create_pin}
\alias{board_pin_get}
\alias{board_remove_pin}
\alias{board_find_pin}
\alias{board_local_storage}
\alias{board_load}
\alias{board_persist}
\title{Custom Boards}
\usage{
board_create_pin(board, path, name, description, type, metadata)

board_pin_get(board, name, details)

board_remove_pin(board, name)

board_find_pin(board, text, ...)

board_local_storage(component)

board_load(board)

board_persist(board)
}
\arguments{
\item{x}{A local file path or an object. Boards must support storing both.}
}
\description{
Family of functions meant to be used to implement custom boards extensions,
not to be used by end users.
}
\keyword{internal}
