% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_load}
\alias{pin_load}
\alias{board_pin_store}
\title{Custom Pins}
\usage{
pin_load(path, ...)

board_pin_store(
  board,
  path,
  name,
  description,
  type,
  metadata,
  extract = TRUE,
  retrieve = TRUE,
  ...
)
}
\arguments{
\item{path}{The path to store.}

\item{...}{Additional parameteres.}

\item{board}{The board to extended, retrieved with \code{board_get()}.}

\item{name}{The name of the pin.}

\item{description}{The text patteren to find a pin.}

\item{type}{The type of pin being stored.}

\item{metadata}{A list containing additional metadata describing the pin.}

\item{retrieve}{Should the pin be retrieved after being created? Defaults to \code{TRUE}.}
}
\description{
Family of functions meant to be used to implement custom pin extensions, not to be used by users.
}
\keyword{internal}
