% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{find_pin}
\alias{find_pin}
\title{Find Pin}
\usage{
find_pin(name = NULL, board = NULL)
}
\arguments{
\item{name}{The name for the pin}

\item{board}{The board where this pin will be placed.}
}
\description{
Find a pin in the active board.
}
