% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{find_pins}
\alias{find_pins}
\title{Find Pins}
\usage{
find_pins(name = NULL, board = default_board())
}
\arguments{
\item{name}{The name for the pin}

\item{board}{The board where this pin will be placed.}
}
\description{
Find pins in the active board.
}
