% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{get_pin}
\alias{get_pin}
\title{Retrieve Pin}
\usage{
get_pin(name, board = NULL, ...)
}
\arguments{
\item{name}{The name of the pin.}

\item{board}{The board where this pin will be retrieved from.}

\item{...}{Additional parameters.}

\item{type}{The type of pin to retrieve.}
}
\description{
Retrieves a named pin from the active board.
}
