% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_local.R
\name{legacy_local}
\alias{legacy_local}
\title{Legacy local board}
\usage{
legacy_local(path = NULL, name = "local", versions = FALSE)
}
\arguments{
\item{name}{An optional name used identify the board. This is no longer
generally needed since you should be passing around an explicit
board object.}

\item{versions}{Should this board be registered with support for versions?}
}
\description{
\code{legacy_local()} powers \code{board_register_local()}, which allows you to
access local pins created in earlier versions of the pins package. For
new pins, we recommend that you transition to \code{\link[=board_local]{board_local()}} which
supports the new pins API.
}
\examples{
# Old api
pin(data.frame(x = 1:3), "test")
pin_get("test")

# New api
board <- board_local()
board \%>\% pin_write(data.frame(x = 1:3), "test")
board \%>\% pin_read("test")
}
\keyword{internal}
