% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Dataset Pin}
\usage{
pin(x = NULL, name = NULL, description = "",
  board = active_board())
}
\arguments{
\item{x}{A dataset to pin or the named pin to retrieve.}

\item{name}{The name for the dataset.}

\item{description}{Optional description for this pin.}

\item{board}{The board where this pin will be placed.}
}
\description{
Pins the given dataset in the active board or retrieves
a named pin from the active board.
}
