% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_browse.R
\name{pin_browse}
\alias{pin_browse}
\title{Browse source of a pin}
\usage{
pin_browse(board, name, version = NULL, ..., cache = FALSE)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_url]{board_url()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{version}{For boards that support versions, this allows you to retrieve
a specific version of a pin. See \verb{vignette("versioning)"} for details.}

\item{...}{Other arguments passed on to methods.}

\item{cache}{If \code{TRUE}, will open the directory on your computer used
to cache pin files.}
}
\description{
\code{pin_browse()} navigates you to the home of a pin, either on the
internet or on your local file system.
}
\examples{
board <- board_temp(versions = TRUE)
board \%>\% pin_write(1:10, "x")
board \%>\% pin_write(1:11, "x")
board \%>\% pin_write(1:12, "x")

board \%>\% pin_browse("x")
}
