% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_cache.R
\name{pin_cache}
\alias{pin_cache}
\title{Cache a pin locally}
\usage{
pin_cache(board, name, version = NULL, ...)
}
\value{
Pin metadata with two extra fields \code{cache_dir} and \code{cache_paths}
}
\description{
This is an internal function, needed primarily for folks who are developing
new board types.
}
\keyword{internal}
