% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_connection.R
\name{pin_connection}
\alias{pin_connection}
\title{Pin Connection}
\usage{
pin_connection(name, initializer = "DBI::dbConnect", driver = NULL,
  ...)
}
\arguments{
\item{name}{The name for this pin.}

\item{initializer}{Function used to connect. Defaults to \code{"DBI::dbConnect"}.}

\item{driver}{The name of the driver with an optional library}

\item{...}{Additional parameters to initialize the connection.}
}
\description{
Describe a connection to be able to \code{pin()} it.
}
