% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_default_name}
\alias{pin_default_name}
\title{Create Pin Name}
\usage{
pin_default_name(x, board)
}
\arguments{
\item{x}{The expression to generate  the pin name from.}

\item{board}{The board to which this name is generating for.}
}
\description{
Creates a pin name from an character expression generated with \code{deparse(substitute(x))}.
}
\keyword{internal}
