% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-upload-download.R
\name{pin_download}
\alias{pin_download}
\alias{pin_upload}
\title{Upload and download files to and from a board}
\usage{
pin_download(path, ...)

pin_upload(board, name, path, desc = NULL, metadata = list(), ...)
}
\description{
This is a lower-level interface than \code{pin_read()} and \code{pin_write()} that
works with paths. You can use this to upload any
}
