% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_exists.R
\name{pin_exists}
\alias{pin_exists}
\title{Does a pin exist in a board?}
\usage{
pin_exists(board, name, ...)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_url]{board_url()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\description{
Does a pin exist in a board?
}
