% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-store-fetch.R
\name{pin_fetch}
\alias{pin_fetch}
\alias{pin_store}
\title{Fetch/store a pin}
\usage{
pin_fetch(board, name, version = NULL, ...)

pin_store(board, name, path, metadata, versioned = NULL, x = NULL, ...)
}
\value{
\code{pin_fetch()} is called primarily for its side-effect of downloading
remote pins into the local cache. It returns the same data as \link{pin_meta}.
\code{pin_store()} is called for its side-effect of uploading a local file
to a remote board. It invisibly returns the \code{board}.
}
\description{
These are low-level functions that power \code{\link[=pin_read]{pin_read()}}, \code{\link[=pin_write]{pin_write()}},
\code{\link[=pin_upload]{pin_upload()}}, and \link{pin_download}. They are needed primarily for folks
developing new board types, and should not generally be called directly.
}
\keyword{internal}
