% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-read-write.R
\name{pin_file_writer}
\alias{pin_file_writer}
\title{Helper to create file-writing function}
\usage{
pin_file_writer(.f, extension)
}
\arguments{
\item{.f}{\code{function} or purrr-style anonymous function that takes
an object and a path, then writes the object to the path.}

\item{extension}{\code{character} extension for the file to be written.}
}
\value{
\code{function} to write a pin file:
\itemize{
\item takes an object to be written, and a name for the pin
\item writes the object to a temporary path, using \code{.f} and \code{extension}
\item returns the path, invisibly
}
}
\description{
Lower-level functions \code{\link[=pin_download]{pin_download()}} and \code{\link[=pin_upload]{pin_upload()}} can be used to
read and write custom-formatted file pins. You can use this function to build
a file-writer for a given format.
}
\examples{
arrow_pin_writer <- pin_file_writer(
  ~arrow::write_feather(.x, .y, compression = "uncompressed"),
  extension = "arrow"
)

board <- board_temp()
pin_upload(board, arrow_pin_writer(mtcars))
pin_write(board, mtcars, type = arrow_pin_writer)
}
