% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_find}
\alias{pin_find}
\title{Find Pin}
\usage{
pin_find(text = NULL, board = NULL, ...)
}
\arguments{
\item{text}{The text to find in the pin description or name.}

\item{board}{The board name used to find the pin.}

\item{...}{Additional parameters.}
}
\description{
Find a pin in any board registered using \code{use_board()} or
\code{board_register()}.
}
