% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_get}
\alias{pin_get}
\title{Retrieve Pin}
\usage{
pin_get(name, board = NULL, ...)
}
\arguments{
\item{name}{The name of the pin.}

\item{board}{The board where this pin will be retrieved from.}

\item{...}{Additional parameters.}
}
\description{
Retrieves a pin by name from the local or given board.
}
