% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_load}
\alias{pin_load}
\title{Load Pin}
\usage{
pin_load(path, ...)
}
\arguments{
\item{path}{The file to load as a pin.}

\item{...}{Additional parameters.}
}
\description{
Load a pin from the given file path making use of the pin type.
}
\keyword{internal}
