% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_manifest.R
\name{pin_manifest}
\alias{pin_manifest}
\title{Write manifest file}
\usage{
pin_manifest(board)
}
\arguments{
\item{board}{A pin board, currently only \code{board_folder()} is supported.}
}
\value{
\code{board}, invisibly.
}
\description{
This can be used if you have a \code{pins::board_folder()} that you want to
serve as a web-site such that others can consume using \code{pins::board_url()}.
}
\details{
This function is called for the side-effect of writing a manifest file,
\code{pins.txt}, to the \code{boards}'s root-directory.
}
\examples{
board <- board_temp()
pin_write(board, mtcars, "mtcars-csv", type = "csv")
pin_write(board, mtcars, "mtcars-json", type = "json")

pin_manifest(board)
fs::path(board$path, "pins.txt") \%>\% readLines() \%>\% cat(sep = "\n")
}
