% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_remove}
\alias{pin_remove}
\title{Remove Pin}
\usage{
pin_remove(name, board = NULL)
}
\arguments{
\item{name}{The name for the dataset.}

\item{board}{The board where this pin will be placed.}
}
\description{
Unpins the given named pin from the active board.
}
