% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_versions.R
\name{pin_versions}
\alias{pin_versions}
\title{Pin versions}
\usage{
pin_versions(board, name, full = deprecated())
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_github]{board_github()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{full}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A data frame with at least a \code{version} column. Some boards may
provided additional data.
}
\description{
List versions available for a pin. See \verb{vignette("versioning)"} for details.
}
\examples{
board <- board_temp(versions = TRUE)

board \%>\% pin_write(data.frame(x = 1:5), name = "df")
board \%>\% pin_write(data.frame(x = 2:6), name = "df")
board \%>\% pin_write(data.frame(x = 3:7), name = "df")

# pin_read() returns the latest version by default
board \%>\% pin_read("df")

# but you can return earlier versions if needed
board \%>\% pin_versions("df")
board \%>\% pin_read("df", version = "91e5f7f417660e3e")
}
