% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins-package.R
\docType{package}
\name{pins-package}
\alias{pins}
\alias{pins-package}
\title{pins: Pin, Discover and Share Resources}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Pin remote resources into a local cache to work
    offline, improve speed and avoid recomputing; discover and share
    resources in local folders, 'GitHub', 'Kaggle' or 'RStudio Connect'.
    Resources can be anything from 'CSV', 'JSON', or image files to
    arbitrary R objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/pins}
  \item Report bugs at \url{https://github.com/rstudio/pins/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Javier Luraschi \email{jluraschi@gmail.com}
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
\keyword{internal}
