% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pins_metadata_create}
\alias{pins_metadata_create}
\title{Pin Extension Metadata}
\usage{
pins_metadata_create(rows, cols)
}
\arguments{
\item{name}{The name of the pin.}

\item{board}{The board where this pin will be retrieved from.}

\item{...}{Additional parameters.}
}
\description{
Creates metadata for a pin when extending boards.
}
\keyword{internal}
