% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{preview_pin}
\alias{preview_pin}
\title{Preview Pin}
\usage{
preview_pin(name, board = NULL, ...)
}
\arguments{
\item{name}{The name of the pin.}

\item{board}{The board where this pin will be retrieved from.}

\item{...}{Additional parameters.}
}
\description{
Previews a named pin from the active board.
}
\keyword{internal}
