% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_azure.R, R/board_ms365.R,
%   R/board_rsconnect.R, R/board_s3.R, R/required_pkgs.R
\name{required_pkgs.pins_board_azure}
\alias{required_pkgs.pins_board_azure}
\alias{required_pkgs.pins_board_ms365}
\alias{required_pkgs.pins_board_rsconnect}
\alias{required_pkgs.pins_board_s3}
\alias{required_pkgs.pins_board}
\title{Determine required packages for a pins board}
\usage{
\method{required_pkgs}{pins_board_azure}(x, ...)

\method{required_pkgs}{pins_board_ms365}(x, ...)

\method{required_pkgs}{pins_board_rsconnect}(x, ...)

\method{required_pkgs}{pins_board_s3}(x, ...)

\method{required_pkgs}{pins_board}(x, ...)
}
\arguments{
\item{x}{A pin board}

\item{...}{Not used.}
}
\value{
A character vector of package names required to use the board.
}
\description{
Determine required packages for a pins board
}
\examples{
required_pkgs(board_temp())

}
