% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{unpin}
\alias{unpin}
\title{Unpin Dataset}
\usage{
unpin(name, board = active_board())
}
\arguments{
\item{name}{The name for the dataset.}

\item{board}{The board where this pin will be placed.}
}
\description{
Unpins the given named pin from the active board.
}
