% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{use_board}
\alias{use_board}
\title{Use Board}
\usage{
use_board(name, ...)
}
\arguments{
\item{name}{The name of the board to activate.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Defines which board to use, defaults to a board storing data
locally under a \code{~/pins} folder.
}
