% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_rsconnect.R
\name{clear_connect_cache}
\alias{clear_connect_cache}
\alias{use_connect_cache}
\title{Detect whether to use content and user caches on RStudio Connect}
\usage{
clear_connect_cache(board, cache = c("content-cache.yml", "users-cache.yml"))

use_connect_cache()
}
\arguments{
\item{board}{A pin board created by \code{\link[=board_rsconnect]{board_rsconnect()}}.}

\item{cache}{Filename for either the content or user cache.}
}
\value{
\code{use_connect_cache()} returns a logical, \code{TRUE} or \code{FALSE}.
\code{clear_connect_cache()} returns the deleted path, invisibly.
}
\description{
Sometimes the content and user caches for RStudio Connect can get out of
sync with the server. While we work on solving this problem, the function
\code{use_connect_cache()} helps you by detecting whether or not to use these
caches. You can manually delete a broken cache via \code{clear_connect_cache()}.
}
\details{
The \code{use_connect_cache()} function uses the following method for
detecting whether or not to use the content and user caches for Connect:
\enumerate{
\item If the \code{pins_connect_cache} option is set to \code{FALSE}, \code{FALSE} is returned.
\item If the \code{PINS_CONNECT_CACHE} environment variable is set to \code{"false"},
\code{"FALSE"}, or \code{"False"}, \code{FALSE} is returned
\item In all other cases (including empty or \code{NULL} values for either of the
above), \code{TRUE} is returned
}

For help deleting a broken cache, you can use the helper function
\code{pins::clear_connect_cache(board, "content-cache.yml")} or
\code{pins::clear_connect_cache(board, "user-cache.yml")} for your Connect \code{board}.
}
\examples{
use_connect_cache()

}
\seealso{
\code{\link[=board_rsconnect]{board_rsconnect()}}
}
