% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{write_board_manifest}
\alias{write_board_manifest}
\title{Write board manifest file to board's root directory}
\usage{
write_board_manifest(board, ...)
}
\arguments{
\item{board}{A pin board that is \emph{not} read-only.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\value{
The board, invisibly
}
\description{
A board manifest file contains a named list of pins, where the values
identify pin-version directories.
This can be useful for a board built using, for example,
\code{\link[=board_folder]{board_folder()}} or \code{\link[=board_s3]{board_s3()}}, then served as a website,
such that others can consume using \code{\link[=board_url]{board_url()}}.
}
\details{
This function is not supported for read-only boards.
It is called for the side-effect of writing a manifest file,
\verb{_pins.yaml}, to the root directory of the  \code{board}.
}
\examples{
board <- board_temp()
pin_write(board, mtcars, "mtcars-csv", type = "csv")
pin_write(board, mtcars, "mtcars-json", type = "json")

write_board_manifest(board)

# see the manifest's format:
fs::path(board$path, "_pins.yaml") \%>\% readLines() \%>\% cat(sep = "\n")

}
