% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{write_board_manifest_yaml}
\alias{write_board_manifest_yaml}
\title{Write a manifest YAML file for a board}
\usage{
write_board_manifest_yaml(board, manifest, ...)
}
\arguments{
\item{board}{A pin board that is \emph{not} read-only.}

\item{manifest}{Contents to be written to the manifest file, as a list.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\value{
\code{write_board_manifest_yaml()} is called for its side-effect of
writing a manifest YAML file.
}
\description{
This is a low-level function that powers \code{\link[=write_board_manifest]{write_board_manifest()}}. It is
needed primarily for folks developing new board types, and should not
generally be called directly.
}
\keyword{internal}
