% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\alias{use_board}
\alias{active_board}
\title{Use Board}
\usage{
use_board(...)

active_board()
}
\arguments{
\item{...}{A list of boards to use, defaults to \code{"local"}.}
}
\description{
Defines which board to use, defaults to a board storing data
locally under a \code{~/pinboard} folder.
}
