% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_viewer.R
\name{board_connect}
\alias{board_connect}
\alias{board_disconnect}
\title{Connect to/disconnect from a board}
\usage{
board_connect(board, code, ...)

board_disconnect(name, ...)
}
\arguments{
\item{board}{The name of the board to activate.}

\item{code}{The code being used to registere this board.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Activate/deactivate RStudio's connection pane
}
\keyword{internal}
