% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_connect}
\alias{board_connect}
\title{Connect to Board}
\usage{
board_connect(board, code, ...)
}
\arguments{
\item{board}{The name of the board to activate.}

\item{code}{The code being used to registere this board.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Connects to a board to activate RStudio's connection pane, when available.
}
\keyword{internal}
