% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_connect_url.R
\name{board_connect_url}
\alias{board_connect_url}
\title{Use a vector of Posit Connect vanity URLs as a board}
\usage{
board_connect_url(
  vanity_urls,
  auth = c("auto", "manual", "envvar", "rsconnect"),
  server = NULL,
  account = NULL,
  key = NULL,
  cache = NULL,
  use_cache_on_failure = is_interactive()
)
}
\arguments{
\item{vanity_urls}{A named character vector of
\href{https://docs.posit.co/connect/user/content-settings/#custom-url}{Connect vanity URLs}.
This board is read only, and the best way to write to a pin on Connect is
\code{\link[=board_connect]{board_connect()}}.}

\item{auth}{There are three ways to authenticate:
\itemize{
\item \code{auth = "manual"} uses arguments \code{server} and \code{key}.
\item \code{auth = "envvar"} uses environment variables \code{CONNECT_API_KEY}
and \code{CONNECT_SERVER}.
\item \code{auth = "rsconnect"} uses servers registered with the rsconnect
package (filtered by \code{server} and \code{account}, if provided)
}

The default, \code{auth = "auto"}, automatically picks between the three options,
using \code{"manual"} if \code{server} and \code{key} are provided, \code{"envvar"} if both
environment variables are set, and \code{"rsconnect"} otherwise.}

\item{server}{For \code{auth = "manual"} or \code{auth = 'envvar'}, the full url to the server,
like \verb{http://server.posit.co/rsc} or \verb{https://connect.posit.co/}.
For \code{auth = 'rsconnect'} a host name used to disambiguate Connect accounts,
like \code{server.posit.co} or \code{connect.posit.co}.}

\item{account}{A user name used to disambiguate multiple Connect accounts.}

\item{key}{The Posit Connect API key.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{use_cache_on_failure}{If the pin fails to download, is it ok to
use the last cached version? Defaults to \code{is_interactive()} so you'll
be robust to poor internet connectivity when exploring interactively,
but you'll get clear errors when the code is deployed.}
}
\description{
\code{board_connect_url()} lets you build up a board from individual
\href{https://docs.posit.co/connect/user/content-settings/#custom-url}{vanity urls}.
To use a Posit Connect board like this one or \code{\link[=board_connect]{board_connect()}}, you need to
first authenticate. The easiest way to do so is by launching \strong{Tools} -
\strong{Global Options} - \strong{Publishing} - \strong{Connect}, and follow the
instructions.

\code{board_connect_url()} is read only, and does not support either the use of a
\link[=write_board_manifest]{manifest file} or versioning.
}
\examples{
\dontshow{if (interactive() || identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
board <- board_connect_url(c(
    my_vanity_url_pin = "https://colorado.posit.co/rsc/great-numbers/"
))

board \%>\% pin_read("my_vanity_url_pin")
\dontshow{\}) # examplesIf}
}
\seealso{
Other boards: 
\code{\link{board_connect}()},
\code{\link{board_folder}()},
\code{\link{board_url}()}
}
\concept{boards}
