% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_kaggle.R
\name{board_kaggle}
\alias{board_kaggle}
\title{Use a Kaggle board}
\usage{
board_kaggle(token = NULL, name = "kaggle", ...)
}
\arguments{
\item{token}{The Kaggle token as a path to the \code{kaggle.json} file, can
be \code{NULL} if the \verb{~/.kaggle/kaggle.json} file already exists.}
}
\description{
To use a Kaggle board, you need to first download a token file from
\url{https://www.kaggle.com/me/account}.
}
\section{Sharing}{
When working in teams, you might want to share your pins with others. For
You can do by adding users or making the dataset public on Kaggle's website.

Once you share with specific users, they can follow the same steps to
register a Kaggle board which allows them to download and upload pins
}

\examples{
\dontrun{
# the following example requires a Kaggle API token
board <- board_kaggle(token = "path/to/kaggle.json")

pin_find("crowdflower", board = board)

# names starting with c/ are competitions
pin_get("c/crowdflower-weather-twitter", board = board)
}
}
\keyword{internal}
