% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_local.R
\name{board_local}
\alias{board_local}
\title{Register Local Board}
\usage{
board_local(cache = board_cache_path(), name = "local", versions = FALSE)
}
\arguments{
\item{cache}{The folder where pins will be read/written.}

\item{name}{An optional name used identify the board.}

\item{versions}{Should this board be registered with support for versions?}
}
\description{
Use a local folder as a board.
}
\examples{
# session-specific local board
board <- board_local(cache = tempfile())
}
\seealso{
Other boards: 
\code{\link{board_azure}()},
\code{\link{board_github}()},
\code{\link{board_rsconnect}()},
\code{\link{board_s3}()}
}
\concept{boards}
