% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_ms365.R
\name{board_ms365}
\alias{board_ms365}
\alias{board_sharepoint}
\alias{board_onedrive}
\title{Use a OneDrive or Sharepoint document library as a board}
\usage{
board_ms365(drive, path, versioned = TRUE, cache = NULL)

board_sharepoint(drive, path, versioned = TRUE, cache = NULL)

board_onedrive(drive, path, versioned = TRUE, cache = NULL)
}
\arguments{
\item{drive}{A OneDrive or SharePoint document library object, as obtained from Microsoft365. See the examples.}

\item{path}{Path to directory to store pins. Will be created if it
doesn't already exist.}

\item{versioned}{Should this board be registered with support for versions?}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}
}
\description{
Pin data to a folder in Onedrive or a SharePoint Online document library
using the Microsoft365R package.
}
\examples{
\dontrun{
# a board in your personal OneDrive
od <- get_personal_onedrive()
odboard <- board_onedrive(od, "myboard")
odboard \%>\% pin_write(iris)

# a board in OneDrive for Business
odb <- get_business_onedrive(tenant="mytenant")
odbboard <- board_onedrive(odb, "myproject/board")

# a board in a SharePoint Online document library
sp <- get_sharepoint_site("my site", tenant="mytenant")
doclib <- sp$get_drive()
spboard <- board_sharepoint(doclib, "general/project1/board")
}
}
