% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_ms365.R
\name{board_ms365}
\alias{board_ms365}
\alias{board_sharepoint}
\alias{board_onedrive}
\title{Use a OneDrive or Sharepoint document library as a board}
\usage{
board_ms365(
  drive,
  path,
  versioned = TRUE,
  cache = NULL,
  delete_by_item = FALSE
)

board_sharepoint(
  drive,
  path,
  versioned = TRUE,
  cache = NULL,
  delete_by_item = FALSE
)

board_onedrive(
  drive,
  path,
  versioned = TRUE,
  cache = NULL,
  delete_by_item = FALSE
)
}
\arguments{
\item{drive}{A OneDrive or SharePoint document library object, as obtained
from Microsoft365.}

\item{path}{Path to directory to store pins. Will be created if it
doesn't already exist.}

\item{versioned}{Should this board be registered with support for versions?}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{delete_by_item}{Whether to handle folder deletions on an item-by-item
basis, rather than deleting the entire folder at once. You may need to set
this to \code{TRUE} for a board in SharePoint Online or OneDrive for Business,
due to document protection policies that prohibit deleting non-empty
folders.}
}
\description{
Pin data to a folder in Onedrive or a SharePoint Online document library
using the Microsoft365R package.
}
\examples{
\dontrun{
# A board in your personal OneDrive
od <- Microsoft365R::get_personal_onedrive()
board <- board_onedrive(od, "myboard")
board \%>\% pin_write(iris)

# A board in OneDrive for Business
odb <- Microsoft365R::get_business_onedrive(tenant = "mytenant")
board <- board_onedrive(odb, "myproject/board")

# A board in a SharePoint Online document library
sp <- Microsoft365R::get_sharepoint_site("my site", tenant = "mytenant")
doclib <- sp$get_drive()
board <- board_sharepoint(doclib, "general/project1/board")
}
}
