% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registry.R
\name{board_register}
\alias{board_register}
\alias{board_deregister}
\alias{board_default}
\alias{board_list}
\title{Board registry}
\usage{
board_register(
  board,
  name = NULL,
  cache = board_cache_path(),
  versions = NULL,
  ...
)

board_deregister(name, ...)

board_default()

board_list()
}
\arguments{
\item{board}{The name of the board to register.}

\item{name}{An optional name to identify this board, defaults to the board name.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{versions}{Should this board be registered with support for versions?}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Prior to pins 1.0.0, boards were managed using a named registry, with
\code{\link[=pin]{pin()}} and \code{\link[=pin_get]{pin_get()}} defaulting to a local board. Now, you must
explicitly supply a \code{board} that you've created using \code{\link[=board_local]{board_local()}},
\code{\link[=board_rsconnect]{board_rsconnect()}} etc.
}
\examples{
# previously
board_register_local("myboard", cache = tempfile())
pin(mtcars, board = "myboard")
pin_get("mtcars", board = "myboard")

# now
board <- board_local(cache = tempfile())
pin(mtcars, board = board)
pin_get("mtcars", board = board)
}
\keyword{internal}
