% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registry.R
\name{board_register}
\alias{board_register}
\alias{board_register_azure}
\alias{board_register_datatxt}
\alias{board_register_dospace}
\alias{board_register_gcloud}
\alias{board_register_github}
\alias{board_register_local}
\alias{board_register_kaggle}
\alias{board_register_rsconnect}
\alias{board_register_s3}
\alias{board_deregister}
\alias{board_default}
\alias{board_list}
\title{Board registry}
\usage{
board_register(
  board,
  name = NULL,
  cache = board_cache_path(),
  versions = NULL,
  ...
)

board_register_azure(
  name = "azure",
  container = Sys.getenv("AZURE_STORAGE_CONTAINER"),
  account = Sys.getenv("AZURE_STORAGE_ACCOUNT"),
  key = Sys.getenv("AZURE_STORAGE_KEY"),
  cache = board_cache_path(),
  path = NULL,
  ...
)

board_register_datatxt(
  url,
  name = NULL,
  headers = NULL,
  cache = board_cache_path(),
  ...
)

board_register_dospace(
  name = "dospace",
  space = Sys.getenv("DO_SPACE"),
  key = Sys.getenv("DO_ACCESS_KEY_ID"),
  secret = Sys.getenv("DO_SECRET_ACCESS_KEY"),
  datacenter = Sys.getenv("DO_DATACENTER"),
  cache = board_cache_path(),
  host = "digitaloceanspaces.com",
  path = NULL,
  ...
)

board_register_gcloud(
  name = "gcloud",
  bucket = Sys.getenv("GCLOUD_STORAGE_BUCKET"),
  token = NULL,
  cache = board_cache_path(),
  path = NULL,
  ...
)

board_register_github(
  name = "github",
  repo = NULL,
  branch = NULL,
  token = NULL,
  path = "",
  host = "https://api.github.com",
  cache = board_cache_path(),
  ...
)

board_register_local(name = "local", cache = board_cache_path(), ...)

board_register_kaggle(
  name = "kaggle",
  token = NULL,
  cache = board_cache_path(),
  ...
)

board_register_rsconnect(
  name = "rsconnect",
  server = NULL,
  account = NULL,
  key = NULL,
  output_files = FALSE,
  cache = board_cache_path(),
  ...
)

board_register_s3(
  name = "s3",
  bucket = Sys.getenv("AWS_BUCKET"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  cache = board_cache_path(),
  host = "s3.amazonaws.com",
  region = NULL,
  path = NULL,
  ...
)

board_deregister(name, ...)

board_default()

board_list()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Prior to pins 1.0.0, boards were managed using a named registry, with
\code{\link[=pin]{pin()}} and \code{\link[=pin_get]{pin_get()}} defaulting to a local board. Now, you must
explicitly supply a \code{board} that you've created using \code{\link[=board_local]{board_local()}},
\code{\link[=board_rsconnect]{board_rsconnect()}} etc.
}
\examples{
# previously
board_register_local("myboard", cache = tempfile())
pin(mtcars, board = "myboard")
pin_get("mtcars", board = "myboard")

# now
board <- board_local(cache = tempfile())
pin(mtcars, board = board)
pin_get("mtcars", board = board)
}
\keyword{internal}
