% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board_register}
\alias{board_register}
\title{Register Board}
\usage{
board_register(name, ...)
}
\arguments{
\item{name}{The name of the board to activate.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Registers a board, useful to add sources to \code{pin_find()}.
}
