% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registration.R
\name{board_register_datatxt}
\alias{board_register_datatxt}
\title{Register Data TXT Board}
\usage{
board_register_datatxt(name, url)
}
\arguments{
\item{name}{The name for this board, usually the domain name of the website.}

\item{url}{Path to the \code{data.txt} file or path containing it.}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register as a board a website describing resources with a \code{data.txt} file.
}
\examples{

# register website board using datatxt file
board_register_datatxt(name = "txtexample", url = "https://datatxt.org/data.txt")

# find pins
pin_find(board = "txtexample")

}
\seealso{
board_register
}
