% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registration.R
\name{board_register_github}
\alias{board_register_github}
\title{Register GitHub Board}
\usage{
board_register_github(name = "github", repo = NULL,
  branch = "master", token = NULL, path = "")
}
\arguments{
\item{name}{Optional name for this board, defaults to 'github'.}

\item{repo}{The GitHub repository formatted as 'owner/repo', can be
\code{NULL} if the \code{GITHUB_PAT} environment variable is set.}

\item{branch}{The branch to use when commiting pins.}

\item{token}{Token to use when \code{GITHUB_PAT} is not specified.}

\item{path}{The subdirectory in the repo where the pins will be stored.}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register a GitHub repo as a board.
}
\examples{
\dontrun{
# the following example requires a GitHub API key
board_register_github(repo = "owner/repo")
}
}
\seealso{
board_register
}
