% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registration.R
\name{board_register_kaggle}
\alias{board_register_kaggle}
\title{Register Kaggle Board}
\usage{
board_register_kaggle(token, name = "kaggle", overwrite = FALSE)
}
\arguments{
\item{token}{The Kaggle token as a path to the \code{kaggle.json} file.}

\item{name}{Optional name for this board, defaults to 'kaggle'.}

\item{overwrite}{Should \code{~/.kaggle/kaggle.json} be overriden?}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register Kaggle as a board.
}
\examples{
\dontrun{
board_register_kaggle(token = "path/to/kaggle.json")
}

}
\seealso{
board_register
}
