% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_kaggle.R
\name{board_register_kaggle}
\alias{board_register_kaggle}
\alias{board_kaggle}
\title{Register Kaggle Board}
\usage{
board_register_kaggle(
  name = "kaggle",
  token = NULL,
  cache = board_cache_path(),
  ...
)

board_kaggle(name, token = NULL, ...)
}
\arguments{
\item{name}{Optional name for this board, defaults to 'kaggle'.}

\item{token}{The Kaggle token as a path to the \code{kaggle.json} file, can
be \code{NULL} if the \verb{~/.kaggle/kaggle.json} file already exists.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
To use a Kaggle board, you need to first download a token file from
\url{https://www.kaggle.com/me/account}.
}
\section{Sharing}{
When working in teams, you might want to share your pins with others. For
You can do by adding users or making the dataset public on Kaggle's website.

Once you share with specific users, they can follow the same steps to
register a Kaggle board which allows them to download and upload pins
}

\examples{
\dontrun{
# the following example requires a Kaggle API token
board_register_kaggle(token = "path/to/kaggle.json")

pin_find("crowdflower", board = "kaggle")

# names starting with c/ are competitions
pin_get("c/crowdflower-weather-twitter", board = "kaggle")
}
}
\seealso{
board_register
}
\keyword{internal}
