% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_local.R
\name{board_register_local}
\alias{board_register_local}
\alias{board_local}
\title{Register Local Board}
\usage{
board_register_local(name = "local", cache = board_cache_path(), ...)

board_local(cache = board_cache_path(), name = "local", versions = FALSE)
}
\arguments{
\item{name}{An optional name to identify this board, defaults to the board name.}

\item{cache}{The folder where pins will be read/written.}

\item{...}{Additional parameters required to initialize a particular board.}

\item{versions}{Should this board be registered with support for versions?}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register a local folder as a board.
}
\examples{
# session-specific local board
board <- board_local(cache = tempfile())
}
\seealso{
board_register

Other boards: 
\code{\link{board_register_azure}()},
\code{\link{board_register_github}()},
\code{\link{board_register_rsconnect}()},
\code{\link{board_register_s3}()}
}
\concept{boards}
