% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_registration.R
\name{board_register_local}
\alias{board_register_local}
\title{Register Local Board}
\usage{
board_register_local(name = "local", cache = board_cache_path(), ...)
}
\arguments{
\item{name}{Optional name for this board, defaults to 'local'.}

\item{cache}{The local folder to use as a cache, defaults to \code{board_cache_path()}.}

\item{...}{Additional parameters required to initialize a particular board.}
}
\description{
Wrapper with explicit parameters over \code{board_register()} to
register a local folder as a board.
}
\examples{
# register local board using a temp folder
board_register_local(cache = tempfile())
}
\seealso{
board_register
}
