% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_test.R
\name{board_test}
\alias{board_test}
\title{Test Board}
\usage{
board_test(board, exclude = list())
}
\arguments{
\item{exclude}{Names of tests to exclude form test}

\item{name}{The name of the board to test}
}
\description{
Tests a particular board, useful only when creating new boards.
}
\keyword{internal}
