% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_test.R
\name{board_test}
\alias{board_test}
\title{Test Board}
\usage{
board_test(board, exclude = list(), suite = c("default", "versions"))
}
\arguments{
\item{board}{The name of the board to test.}

\item{exclude}{Names of tests to exclude form test.}

\item{suite}{The test suite to run, currently only \code{"versions"} or
\code{default} are supported.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Tests a particular board, useful only when creating new boards.
}
\keyword{internal}
