% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_test.R
\name{board_test}
\alias{board_test}
\title{Test Board}
\usage{
board_test(board, exclude = list(), suite = c("default", "versions"))
}
\arguments{
\item{exclude}{Names of tests to exclude form test.}

\item{suite}{The test suite to run, currently only \code{"versions"} or
\code{default} are supported.}

\item{name}{The name of the board to test.}
}
\description{
Tests a particular board, useful only when creating new boards.
}
\keyword{internal}
