% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_url.R
\name{board_url}
\alias{board_url}
\title{Use a vector of URLs as a board}
\usage{
board_url(urls, cache = NULL, use_cache_on_failure = is_interactive())
}
\arguments{
\item{urls}{identify available pins (see details):
\itemize{
\item unnamed character scalar: URL to a \link[=pin_manifest]{manifest file}.
\item named list: supports versioning, values are URLs to version-directories.
\item named character vector: does not support versioning, values are
URLs to data.
}}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{use_cache_on_failure}{If the pin fails to download, is it ok to
use the last cached version? Defaults to \code{is_interactive()} so you'll
be robust to poor internet connectivity when exploring interactively,
but you'll get clear errors when the code is deployed.}
}
\description{
\code{board_url()} lets you build up a board from individual urls. This is
useful because \code{\link[=pin_download]{pin_download()}} and \code{\link[=pin_read]{pin_read()}} will be cached - they'll
only re-download the data if it's changed from the last time you downloaded
it (using the tools of
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Caching}{HTTP caching}).
You'll also be protected from the vagaries of the internet; if a fresh
download fails, you'll get the previously cached result with a warning.

\code{board_url()} is read only.
}
\details{
The nature of the board depends on the type of the \code{urls} argument:
\itemize{
\item unnamed character scalar, i.e. single string:
URL to a \link[=pin_manifest]{manifest file}.
If the URL ends in a \code{/}, \code{board_url()} will look for a \code{pins.txt}
containing the manifest. If manifest-file parses to a named list,
versioning is supported. If it parses to a named character vector,
the board will not support versioning.
\item named list, values are character vectors of URLs, each element of the
vector refers to a version of the particular pin. If a URL ends in a \code{/},
\code{board_url()} will look for a \code{data.txt} that provides metadata.
Versioning is supported.
\item named character vector of URLs: If the URL ends in a \code{/},
\code{board_url()} will look for a \code{data.txt} that provides metadata. The
easiest way to generate this file is to upload a pin directory created by
\code{\link[=board_folder]{board_folder()}}. Versioning is not supported.
}
}
\examples{
github_raw <- function(x) paste0("https://raw.githubusercontent.com/", x)

board <- board_url(c(
  files = github_raw("rstudio/pins-r/master/tests/testthat/pin-files/"),
  rds = github_raw("rstudio/pins-r/master/tests/testthat/pin-rds/"),
  raw = github_raw("rstudio/pins-r/master/tests/testthat/pin-files/first.txt")
))

board \%>\% pin_read("rds")
board \%>\% pin_browse("rds", local = TRUE)

board \%>\% pin_download("files")
board \%>\% pin_download("raw")

# use manifest file - TODO: update address
board_manifest <-
  github_raw("ijlyttle/pinsManifest/main/tests/testthat/pins/") \%>\%
  board_url()

board_manifest \%>\% pin_list()
board_manifest \%>\% pin_versions("mtcars-json")
board_manifest \%>\% pin_read("mtcars-json") \%>\% head()
}
\seealso{
Other boards: 
\code{\link{board_folder}()},
\code{\link{board_rsconnect}()}
}
\concept{boards}
