% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_url.R
\name{board_url}
\alias{board_url}
\title{Build a pin board from a vector of URLs}
\usage{
board_url(urls, cache = NULL)
}
\arguments{
\item{urls}{A named character vector of URLs If the URL ends in a \code{/},
\code{board_url} will look for a \code{data.txt} that provides metadata.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}
}
\description{
\code{board_url()} lets you build up a board from individual urls. This is
useful because \code{\link[=pin_download]{pin_download()}} and \code{\link[=pin_get]{pin_get()}} will be cached - they'll
only re-download the data if it's changed from the last time you downloaded
it (using the tools of
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Caching}{HTTP caching}).

\code{board_url()} is read only and does not currently support versions.
}
\examples{
github_raw <- "https://raw.githubusercontent.com/"
board <- board_url(c(
  files = paste0(github_raw, "rstudio/pins/master/tests/testthat/pin-files/"),
  rds = paste0(github_raw, "rstudio/pins/master/tests/testthat/pin-rds/"),
  raw = paste0(github_raw, "rstudio/pins/master/tests/testthat/pin-files/first.txt")
))

board \%>\% pin_read("rds")

board \%>\% pin_download("files")
board \%>\% pin_download("raw")
}
