% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_extensions.R
\name{board_local_storage}
\alias{board_local_storage}
\title{Custom Boards Utilities}
\usage{
board_local_storage(component = "local")
}
\description{
A set of utilities used when implementing custom boards.
}
\keyword{internal}
