% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_extensions.R
\name{board_pin_create}
\alias{board_pin_create}
\alias{board_initialize}
\alias{board_pin_get}
\alias{board_pin_remove}
\alias{board_pin_find}
\alias{board_load}
\alias{board_persist}
\alias{board_pin_store}
\alias{board_browse}
\title{Custom Boards}
\usage{
board_pin_create(board, path, name, metadata, ...)

board_initialize(board, ...)

board_pin_get(board, name, details, ...)

board_pin_remove(board, name, ...)

board_pin_find(board, text, ...)

board_load(board, ...)

board_persist(board, ...)

board_pin_store(board, path, name, description, type, metadata, ...)

board_browse(board, ...)
}
\arguments{
\item{board}{The board to extended, retrieved with \code{board_get()}.}

\item{path}{The path to store.}

\item{name}{The name of the pin.}

\item{...}{Additional parameteres.}

\item{text}{The text patteren to find a pin.}
}
\description{
Family of functions meant to be used to implement custom boards extensions,
not to be used by users.
}
