% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_extensions.R
\name{board_pin_create}
\alias{board_pin_create}
\alias{board_initialize}
\alias{board_pin_get}
\alias{board_pin_remove}
\alias{board_pin_find}
\alias{board_local_storage}
\alias{board_load}
\alias{board_persist}
\alias{board_info}
\title{Custom Boards}
\usage{
board_pin_create(board, path, name, description, type, metadata)

board_initialize(board, ...)

board_pin_get(board, name, details)

board_pin_remove(board, name)

board_pin_find(board, text, ...)

board_local_storage(component)

board_load(board)

board_persist(board)

board_info(board)
}
\arguments{
\item{x}{A local file path or an object. Boards must support storing both.}
}
\description{
Family of functions meant to be used to implement custom boards extensions,
not to be used by end users.
}
\keyword{internal}
