% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_extensions.R
\name{board_pin_create}
\alias{board_pin_create}
\alias{board_initialize}
\alias{board_pin_get}
\alias{board_pin_remove}
\alias{board_pin_find}
\alias{board_pin_versions}
\alias{board_browse}
\title{Custom Boards}
\usage{
board_pin_create(board, path, name, metadata, ...)

board_initialize(board, ...)

board_pin_get(board, name, ...)

board_pin_remove(board, name, ...)

board_pin_find(board, text, ...)

board_pin_versions(board, name, ...)

board_browse(board, ...)
}
\arguments{
\item{board}{The board to extend, retrieved with \code{board_get()}.}

\item{path}{The path to store as a pin.}

\item{name}{The name of the pin.}

\item{metadata}{A list of metadata associated with this pin.}

\item{...}{Additional parameteres.}

\item{text}{The text patteren to find a pin.}
}
\description{
Family of functions meant to be used to implement custom boards extensions,
not to be used by users.
}
