% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_board.R
\name{board_pin_create}
\alias{board_pin_create}
\alias{board_initialize}
\alias{board_browse}
\alias{board_desc}
\alias{board_pin_get}
\alias{board_pin_remove}
\alias{board_pin_find}
\alias{board_pin_versions}
\title{Custom Boards}
\usage{
board_pin_create(board, path, name, metadata, ...)

board_initialize(board, ...)

board_browse(board, ...)

board_desc(board, ...)

board_pin_get(board, name, ...)

board_pin_remove(board, name, ...)

board_pin_find(board, text, ...)

board_pin_versions(board, name, ...)
}
\arguments{
\item{board}{The board to extend, retrieved with \code{board_get()}.}

\item{path}{The path to store as a pin.}

\item{name}{The name of the pin.}

\item{metadata}{A list of metadata associated with this pin.}

\item{...}{Additional parameters.}

\item{text}{The text pattern to find a pin.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Family of functions meant to be used to implement custom boards extensions,
not to be used by users.
}
\keyword{internal}
