% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R, R/pin_extensions.R
\name{pin_preview}
\alias{pin_preview}
\alias{pin_load}
\alias{board_pin_store}
\title{Custom Pins}
\usage{
pin_preview(x, board = NULL, ...)

pin_load(path, ...)

board_pin_store(board, path, name, description, type, metadata, ...)
}
\arguments{
\item{board}{The board to extended, retrieved with \code{board_get()}.}

\item{...}{Additional parameteres.}

\item{path}{The path to store.}

\item{name}{The name of the pin.}

\item{description}{The text patteren to find a pin.}

\item{type}{The type of pin being stored.}

\item{metadata}{A list containing additional metadata desecribing the pin.}
}
\description{
Family of functions meant to be used to implement custom pin extensions, not to be used by users.
}
\keyword{internal}
