% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_load}
\alias{pin_load}
\alias{board_pin_store}
\title{Custom Pins}
\usage{
pin_load(path, ...)

board_pin_store(
  board,
  path,
  name,
  pin_metadata,
  extract = TRUE,
  retrieve = TRUE,
  zip = FALSE,
  cache = TRUE,
  metadata = NULL,
  custom_metadata = NULL,
  ...
)
}
\arguments{
\item{path}{The path to store.}

\item{...}{Additional parameteres.}

\item{board}{The board to extended, retrieved with \code{board_get()}.}

\item{name}{The name of the pin.}

\item{pin_metadata}{A list of pin metadata describing the pin. Must contain
\code{type} and \code{description}.}

\item{retrieve}{Should the pin be retrieved after being created? Defaults to \code{TRUE}.}

\item{metadata}{Additional user supplied metadata.}

\item{custom_metadata}{Deprecated. Please use \code{metadata} instead.}
}
\description{
Family of functions meant to be used to implement custom pin extensions, not to be used by users.
}
\keyword{internal}
