% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{new_board}
\alias{new_board}
\alias{new_board_v0}
\alias{new_board_v1}
\title{Create a new board}
\usage{
new_board(board, api, cache, ...)

new_board_v0(board, name, cache = NULL, versions = FALSE, ...)

new_board_v1(board, cache, versioned = FALSE, ...)
}
\arguments{
\item{board}{The name of the board to register.}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{...}{Additional parameters required to initialize a particular board.}

\item{name}{An optional name used identify the board. This is no longer
generally needed since you should be passing around an explicit
board object.}

\item{versions, versioned}{Should this board be registered with support for versions?}
}
\description{
Create a new board
}
\keyword{internal}
