% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Create Pin}
\usage{
pin(x, name = NULL, description = NULL, board = NULL, ...)
}
\arguments{
\item{x}{The dataset or object to pin.}

\item{name}{The name for the dataset or object.}

\item{description}{Optional description for this pin.}

\item{board}{The board where this pin will be placed.}

\item{...}{Additional parameters.}
}
\description{
Pins the given dataset or object in the active board.
}
