% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Pin Dataset}
\usage{
pin(dataset = NULL, name, description = "", board = active_board())
}
\arguments{
\item{dataset}{The dataset to pin.}

\item{name}{The name for the dataset.}

\item{description}{Optional description for this pin.}

\item{board}{The board where this pin will be placed.}
}
\description{
Pins the given dataset in the active board.
}
