% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-upload-download.R
\name{pin_download}
\alias{pin_download}
\alias{pin_upload}
\title{Upload and download files to and from a board}
\usage{
pin_download(board, name, ...)

pin_upload(board, path, name = NULL, desc = NULL, metadata = list(), ...)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_github]{board_github()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{...}{Additional arguments passed on to board pin methods.}

\item{path}{Path to file to upload to \code{board}.}

\item{desc}{A text description of the pin; most important for
shared boards so that others can understand what the pin contains.}

\item{metadata}{A list containing additional metadata to store with the pin.
When retrieving the pin, this will be stored in the \code{user} key, to
avoid potential clashes with the metadata that pins itself uses.}
}
\description{
This is a lower-level interface than \code{pin_read()} and \code{pin_write()} that
you can use to pin any file.
}
\examples{
board <- board_temp()

board \%>\% pin_upload(system.file("CITATION"))
path <- board \%>\% pin_download("CITATION")
path
readLines(path)[1:5]
}
