% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-read-write.R
\name{pin_file_writer}
\alias{pin_file_writer}
\title{Helper to create pin-writing function}
\usage{
pin_file_writer(.f, extension)
}
\arguments{
\item{.f}{\code{function} or purrr-style anonymous function that takes
an object and a path, then writes the object to the path.}

\item{extension}{\code{character} extension for the file to be written.}
}
\value{
\code{function} to write a pin:
\itemize{
\item takes an object to be written, and a name for the pin
\item writes the object to a temporary path, using \code{.f} and \code{extension}
\item returns the path, invisibly
}
}
\description{
Helper to create pin-writing function
}
\examples{
  arrow_pin_writer <- pin_file_writer(
    \(x, path) arrow::write_feather(x, path, compression = "uncompressed"),
    extension = "arrow"
  )
}
