% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_find.R
\name{pin_find}
\alias{pin_find}
\title{Search for pins (legacy API)}
\usage{
pin_find(
  text = NULL,
  board = NULL,
  name = NULL,
  extended = FALSE,
  metadata = FALSE,
  ...
)
}
\arguments{
\item{text}{The text to find in the pin description or name.}

\item{board}{The board name used to find the pin.}

\item{name}{The exact name of the pin to match when searching.}

\item{extended}{Should additional board-specific columns be shown?}

\item{metadata}{Include pin metadata in results?}

\item{...}{Additional parameters.}
}
\description{
Search for pins in legacy boards.
}
\examples{
pin_find("cars")
# ->
board <- board_local()
board \%>\% pin_search("cars")
}
