% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_info.R
\name{pin_info}
\alias{pin_info}
\title{Pin Info}
\usage{
pin_info(
  name,
  board = NULL,
  extended = TRUE,
  metadata = TRUE,
  signature = FALSE,
  ...
)
}
\arguments{
\item{name}{The exact name of the pin to match when searching.}

\item{board}{The board name used to find the pin.}

\item{extended}{Should additional board-specific information be shown?}

\item{metadata}{Should additional pin-specific information be shown?}

\item{signature}{Should a signature to identify this pin be shown?}

\item{...}{Additional parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{pin_info()} is deprecated in favour of \code{\link[=pin_meta]{pin_meta()}}.
}
\examples{
board <- board_temp()
pin(mtcars, board = board)

pin_info("mtcars", board = board)
# ->
board \%>\% pin_meta("mtcars")
}
\keyword{internal}
