% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-meta.R
\name{pin_meta}
\alias{pin_meta}
\title{Retrieve metadata for a pin}
\usage{
pin_meta(board, name, version = NULL, ...)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_github]{board_github()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{version}{For boards that support versions, this allows you to retrieve
a specific version of a pin. See \verb{vignette("versioning)"} for details.}

\item{...}{Additional arguments passed on to methods for a specific board.}
}
\value{
A list.
}
\description{
Pin metadata comes from three sources:
\itemize{
\item Standard metadata added by \code{pin_upload()}/\code{pin_write()}.
\item Metadata supplied by the user, stored in \verb{$user}.
\item Local metadata generated when caching the pin, stored in \verb{$local}.
This includes information like the version of the pin, and the path
its local cache.
}
}
\examples{
b <- board_temp()
b \%>\% pin_write(head(mtcars), "mtcars", metadata = list("Hadley" = TRUE))

# Get the pin
b \%>\% pin_read("mtcars")
# Get its metadata
b \%>\% pin_meta("mtcars")
# Get path to underlying data
b \%>\% pin_download("mtcars")
}
