% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_preview}
\alias{pin_preview}
\title{Preview Pin}
\usage{
pin_preview(x, board = NULL, ...)
}
\arguments{
\item{x}{The pin to preview, retrieved with \code{pin_get()}.}

\item{board}{The board where this pin will be retrieved from.}

\item{...}{Additional parameters.}
}
\description{
Previews a subset of the pin contents, useful to print or display
a subset of the pin contents.
}
\keyword{internal}
