% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_viewer.R
\name{pin_preview}
\alias{pin_preview}
\title{Preview a pin (legacy API)}
\usage{
pin_preview(x, board = NULL, ...)
}
\description{
\code{pin_preview()} is no longer used and only exists for legacy compatibility.
}
\keyword{internal}
