% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_viewer.R
\name{pin_preview}
\alias{pin_preview}
\title{Preview a pin (legacy API)}
\usage{
pin_preview(x, board = NULL, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{pin_preview()} is no longer supported.
}
\keyword{internal}
