% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-read-write.R
\name{pin_read}
\alias{pin_read}
\alias{pin_write}
\title{Read and write pins}
\usage{
pin_read(board, name)

pin_write(board, x, name = NULL, description = NULL, metadata = NULL)
}
\arguments{
\item{board}{A pin board, created by \code{\link[=board_folder]{board_folder()}}, \code{\link[=board_rsconnect]{board_rsconnect()}},
\code{\link[=board_github]{board_github()}} or other other \code{board_} function.}

\item{name}{Pin name.}

\item{x}{An object (typically a data frame) to pin.}

\item{description}{A text description of the pin; most important for
shared boards so that others can understand what the pin contains.}

\item{metadata}{A list containing additional metadata to store with the pins.}
}
\description{
Use \code{pin_write()} to pin an object to board, and \code{pin_read()} to retrieve
it.
}
\examples{
b <- board_temp()

b \%>\% pin_write(mtcars, "mtcars")
b

b \%>\% pin_read("mtcars")
}
