% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_remove}
\alias{pin_remove}
\title{Remove Pin}
\usage{
pin_remove(name, board)
}
\arguments{
\item{name}{The name for the pin.}

\item{board}{The board from where this pin will be removed.}
}
\description{
Unpins the given named pin from the given board.
}
\details{
Notice that some boards do not support deleting pins, this is the case
for the Kaggle board. For these boards, you would manually have to
remote resources using the tools the board provides.
}
\examples{

library(pins)

# define local board
board_register_local(cache = tempfile())

# create mtcars pin
pin(mtcars)

# remove mtcars pin
pin_remove(mtcars, board = "local")
}
