% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_list.R
\name{pin_search}
\alias{pin_search}
\title{Search for pins}
\usage{
pin_search(board, pattern = NULL)
}
\description{
Search for pins
}
\examples{
board <- board_temp()

board \%>\% pin_write(1:5, "x", desc = "Some numbers")
board \%>\% pin_write(letters[c(1, 5, 10, 15, 21)], "y", desc = "My favourite letters")
board \%>\% pin_write(runif(20), "z", desc = "Random numbers")

board \%>\% pin_search()
board \%>\% pin_search("number")
board \%>\% pin_search("letters")
}
