% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins-package.R
\docType{package}
\name{pins-package}
\alias{pins}
\alias{pins-package}
\title{pins: Pin, Discover and Share Resources}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Publish data sets, models, and other R objects, making it easy to share them across projects and with your colleagues. You can pin objects to a variety of "boards", including local folders (to share on a networked drive or with 'DropBox'), 'RStudio' connect, Amazon S3, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pins.rstudio.com/}
  \item \url{https://github.com/rstudio/pins-r}
  \item Report bugs at \url{https://github.com/rstudio/pins-r/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@posit.co} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
  \item Javier Luraschi \email{jluraschi@gmail.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
